Add a flag for skipping the database setup
For example, in the run-mooc-grader container,
when the environment variable `grader_NO_DATABASE` is set to
a non-empty value, then run-mooc-grader skips the database setup
and migrations.

In create-db.sh, the -u parameter was removed from /bin/sh because
the script must not exit when the no-database flag is not set.