Remove verification of the s6 overlay download with gpg
The s6 overlay project stopped providing GPG keys for verifying
the download signatures. The new releases are also not signed
at all. New releases are accompanied with sha256 checksums, but
there are no sha256 checksums for the old releases before v3.0.
Therefore, there is no convenient way to verify the signature
of the downloaded s6 package.

* https://github.com/just-containers/s6-overlay/issues/352#issuecomment-1020259107
* README changed:
  https://github.com/just-containers/s6-overlay/commit/70146fd62744fe679992e70b62dda88305b6856b
* https://github.com/just-containers/s6-overlay#verifying-downloads